+++ src/components/reports/pdftemplate/pages/sitemanagement/index.ts
@@ -1,1 +1,1 @@
-export { SiteManagementReportDownloadButton } from './SiteManagementReport';
+export { SiteManagementReportPDF } from './SiteManagementReport';
